requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
google-generativeai==0.8.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub>=0.26.1
idna==3.10
jiter==0.6.1
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
openai==1.58.1
click>=8.0.0
grpcio==1.60.1
groq==0.18.0
